test: add tests for EditFavoritesPresenter
